# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Monitoring API
host: 'localhost:10005'
basePath: /mon
schemes:
  - http
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
  - text/plain
paths:
...
  '/webHookStatus/{id}':
    x-swagger-router-controller: controller
    get:
      description: Retrieve a webHookStatus
      operationId: webHookStatus
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: id
          in: path
          type: string
          required: true
          description: Id of the object to be viewed
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No results to list with the given parameter set.
        '500':
          description: Internal server error        
...